+++ src/http/parser/response.cpp
@@ -124,8 +124,8 @@ void Response::parseHTTPBody(const std::vector<uint8_t>& data, size_t headerLeng
     if (headers.find("transfer-encoding") != headers.end() && finalTransferEncoding == "chunked") {
         parseChunked(data, headers, length, headerLength, body);
     } else {
-        body.insert(body.end(), data.begin() + (long long) headerLength, data.begin()
-            + (long long) headerLength + (long long) length);
+        body.insert(body.end(), data.begin() + (ssize_t) headerLength, data.begin()
+            + (ssize_t) headerLength + (ssize_t) length);
     }
 }
